complete tests for UnorderedMap
